# API/CertificateOfOrigins.WebApi/.spec/OpenApi/api.openapi.json
{
    "openapi":  "3.1.1",
    "info":  {
                 "title":  "CertificateOfOrigins api",
                 "version":  "1.0.0",
                 "description":  "C14 code-first contract for the api (external / service-to-service) domain. Exported from /openapi/v1.json of the running application, filtered to the api/ prefix, with components.schemas reduced to the transitive $ref closure of those paths. HttpQuery endpoints are absent by construction - ASP.NET does not emit them into the generated document."
             },
    "paths":  {
                  "/api/CertificateOfOrigins/ID/{certificateNumber}":  {
                                                                           "get":  {
                                                                                       "tags":  [
                                                                                                    "CertificateOfOrigins"
                                                                                                ],
                                                                                       "parameters":  [
                                                                                                          {
                                                                                                              "name":  "certificateNumber",
                                                                                                              "in":  "path",
                                                                                                              "required":  true,
                                                                                                              "schema":  {
                                                                                                                             "type":  "string"
                                                                                                                         }
                                                                                                          },
                                                                                                          {
                                                                                                              "name":  "X-API-Version",
                                                                                                              "in":  "header",
                                                                                                              "schema":  {
                                                                                                                             "type":  "string"
                                                                                                                         }
                                                                                                          }
                                                                                                      ],
                                                                                       "responses":  {
                                                                                                         "400":  {
                                                                                                                     "description":  "Bad Request",
                                                                                                                     "content":  {
                                                                                                                                     "application/json":  {
                                                                                                                                                              "schema":  {
                                                                                                                                                                             "$ref":  "#/components/schemas/RestBadRequestResult"
                                                                                                                                                                         }
                                                                                                                                                          }
                                                                                                                                 }
                                                                                                                 },
                                                                                                         "404":  {
                                                                                                                     "description":  "Not Found",
                                                                                                                     "content":  {
                                                                                                                                     "text/plain":  {
                                                                                                                                                        "schema":  {
                                                                                                                                                                       "type":  "string"
                                                                                                                                                                   }
                                                                                                                                                    },
                                                                                                                                     "application/octet-stream":  {
                                                                                                                                                                      "schema":  {
                                                                                                                                                                                     "type":  "string"
                                                                                                                                                                                 }
                                                                                                                                                                  },
                                                                                                                                     "application/json":  {

                                                                                                                                                          }
                                                                                                                                 }
                                                                                                                 },
                                                                                                         "200":  {
                                                                                                                     "description":  "OK",
                                                                                                                     "content":  {
                                                                                                                                     "application/json":  {
                                                                                                                                                              "schema":  {
                                                                                                                                                                             "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                                                                                             "type":  [
                                                                                                                                                                                          "integer",
                                                                                                                                                                                          "string"
                                                                                                                                                                                      ],
                                                                                                                                                                             "format":  "int32"
                                                                                                                                                                         }
                                                                                                                                                          }
                                                                                                                                 }
                                                                                                                 }
                                                                                                     }
                                                                                   }
                                                                       },
                  "/api/CertificateOfOrigins/GoodsItemCerificateDTO":  {
                                                                           "post":  {
                                                                                        "tags":  [
                                                                                                     "CertificateOfOrigins"
                                                                                                 ],
                                                                                        "parameters":  [
                                                                                                           {
                                                                                                               "name":  "X-API-Version",
                                                                                                               "in":  "header",
                                                                                                               "schema":  {
                                                                                                                              "type":  "string"
                                                                                                                          }
                                                                                                           }
                                                                                                       ],
                                                                                        "requestBody":  {
                                                                                                            "content":  {
                                                                                                                            "application/json":  {
                                                                                                                                                     "schema":  {
                                                                                                                                                                    "type":  "array",
                                                                                                                                                                    "items":  {
                                                                                                                                                                                  "$ref":  "#/components/schemas/GoodsItemCerificateDto"
                                                                                                                                                                              }
                                                                                                                                                                }
                                                                                                                                                 },
                                                                                                                            "text/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "type":  "array",
                                                                                                                                                             "items":  {
                                                                                                                                                                           "$ref":  "#/components/schemas/GoodsItemCerificateDto"
                                                                                                                                                                       }
                                                                                                                                                         }
                                                                                                                                          },
                                                                                                                            "application/*+json":  {
                                                                                                                                                       "schema":  {
                                                                                                                                                                      "type":  "array",
                                                                                                                                                                      "items":  {
                                                                                                                                                                                    "$ref":  "#/components/schemas/GoodsItemCerificateDto"
                                                                                                                                                                                }
                                                                                                                                                                  }
                                                                                                                                                   }
                                                                                                                        },
                                                                                                            "required":  true
                                                                                                        },
                                                                                        "responses":  {
                                                                                                          "400":  {
                                                                                                                      "description":  "Bad Request",
                                                                                                                      "content":  {
                                                                                                                                      "application/json":  {
                                                                                                                                                               "schema":  {
                                                                                                                                                                              "$ref":  "#/components/schemas/RestBadRequestResult"
                                                                                                                                                                          }
                                                                                                                                                           }
                                                                                                                                  }
                                                                                                                  },
                                                                                                          "404":  {
                                                                                                                      "description":  "Not Found",
                                                                                                                      "content":  {
                                                                                                                                      "text/plain":  {
                                                                                                                                                         "schema":  {
                                                                                                                                                                        "type":  "string"
                                                                                                                                                                    }
                                                                                                                                                     },
                                                                                                                                      "application/octet-stream":  {
                                                                                                                                                                       "schema":  {
                                                                                                                                                                                      "type":  "string"
                                                                                                                                                                                  }
                                                                                                                                                                   },
                                                                                                                                      "application/json":  {

                                                                                                                                                           }
                                                                                                                                  }
                                                                                                                  },
                                                                                                          "200":  {
                                                                                                                      "description":  "OK",
                                                                                                                      "content":  {
                                                                                                                                      "application/json":  {
                                                                                                                                                               "schema":  {
                                                                                                                                                                              "type":  "array",
                                                                                                                                                                              "items":  {
                                                                                                                                                                                            "$ref":  "#/components/schemas/GoodsItemCerificateDto"
                                                                                                                                                                                        }
                                                                                                                                                                          }
                                                                                                                                                           }
                                                                                                                                  }
                                                                                                                  }
                                                                                                      }
                                                                                    }
                                                                       },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "GoodsItemCerificateDto":  {
                                                                      "type":  "object",
                                                                      "properties":  {
                                                                                         "goodsItemId":  {
                                                                                                             "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                             "type":  [
                                                                                                                          "integer",
                                                                                                                          "string"
                                                                                                                      ],
                                                                                                             "format":  "int32"
                                                                                                         },
                                                                                         "certificateNumber":  {
                                                                                                                   "type":  [
                                                                                                                                "null",
                                                                                                                                "string"
                                                                                                                            ]
                                                                                                               },
                                                                                         "certificateOfOriginId":  {
                                                                                                                       "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                                       "type":  [
                                                                                                                                    "null",
                                                                                                                                    "integer",
                                                                                                                                    "string"
                                                                                                                                ],
                                                                                                                       "format":  "int32"
                                                                                                                   }
                                                                                     }
                                                                  },
...
                                       "RestBadRequestResult":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "type":  {
                                                                                                    "type":  [
                                                                                                                 "null",
                                                                                                                 "string"
                                                                                                             ]
                                                                                                },
                                                                                       "title":  {
                                                                                                     "type":  "string"
                                                                                                 },
                                                                                       "status":  {
                                                                                                      "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                      "type":  [
                                                                                                                   "null",
                                                                                                                   "integer",
                                                                                                                   "string"
                                                                                                               ],
                                                                                                      "format":  "int32"
                                                                                                  },
                                                                                       "traceId":  {
                                                                                                       "type":  [
                                                                                                                    "null",
                                                                                                                    "string"
                                                                                                                ]
                                                                                                   },
                                                                                       "detail":  {
                                                                                                      "type":  [
                                                                                                                   "null",
                                                                                                                   "string"
                                                                                                               ]
                                                                                                  },
                                                                                       "instance":  {
                                                                                                        "type":  [
                                                                                                                     "null",
                                                                                                                     "string"
                                                                                                                 ]
                                                                                                    },
                                                                                       "errors":  {
                                                                                                      "type":  "array",
                                                                                                      "items":  {
                                                                                                                    "$ref":  "#/components/schemas/RestBadRequestError"
                                                                                                                }
                                                                                                  }
                                                                                   }
                                                                },
...
                                   }
                   }
}